+++ Adj/Algebra/Set.hs
@@ -1,1 +1,7 @@
 module Adj.Algebra.Set where
+
+infixr 7 :*:, :+:
+
+data (:*:) l r = l :*: r
+
+data (:+:) l r = This l | That r