+++ Adj/Program/Controlflow/Implementation/State.hs
@@ -1,6 +1,7 @@
 module Adj.Program.Controlflow.Implementation.State where
 
-import Adj.Algebra.Category ((.), (.:), type (-->), Flat (Flat), (:*:) ((:*:)), type (:*:>))
+import Adj.Algebra.Category ((.), (.:), type (-->), Flat (Flat), type (:*:>))
+import Adj.Algebra.Set ((:*:) ((:*:)))
 import Adj.Auxiliary (type (=!?=), FG (FG))
 
 type State state = (-->) state =!?= (:*:>) state