+++ Adj/Program/Primitive/Maybe.hs
@@ -1,6 +1,7 @@
 module Adj.Program.Primitive.Maybe where
 
-import Adj.Algebra.Category (Flat (Flat), (:+:) (This, That), Terminal (Terminal))
+import Adj.Algebra.Category (Flat (Flat), Terminal (Terminal))
+import Adj.Algebra.Set ((:+:) (This, That))
 
 type Maybe = Flat (:+:) Terminal